메모 불러오기 성공

+++ discordMemoBot/src/main/java/www/disbot/jmemo/api/framework/mapper/GeneralMapper.java
@@ -5,7 +5,7 @@ import org.apache.ibatis.annotations.Select;
 
 public interface GeneralMapper {
 	@Select("SELECT FOUND_ROWS()")
-	public long getFoundRows();
+	public int getFoundRows();
 	
 	@Select("SELECT NEXT_MULTI_PK(#{tName}, ${summonCnt})")
 	public String getNextMultiIdConcat(
